app/requirements.txt
Django==4.2
django-filter==23.2
django-import-export==3.2.0
django-tables2==2.5.3
gunicorn==20.1.0
psycopg2-binary==2.9.6
